More Storable Wallets
- Increased wallet private key from a 32bit integer to a 64bit integer

+++ create_tables.sql
@@ -1,7 +1,7 @@
 CREATE EXTENSION IF NOT EXISTS CITEXT;
 
 CREATE TABLE IF NOT EXISTS wallets (
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
 	address CITEXT NOT NULL,
 	CONSTRAINT wallet_address UNIQUE (address),
 	CONSTRAINT wallet_address_length CHECK (LENGTH(address) = 42)
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS stars (
 	is_for_sale BOOLEAN NOT NULL,
 	price_ether NUMERIC(30,18),
 	date_created TIMESTAMP NOT NULL,
-	owner_id INTEGER REFERENCES wallets(id) NOT NULL,
+	owner_id BIGINT REFERENCES wallets(id) NOT NULL,
 	CONSTRAINT star_name UNIQUE (name),
 	CONSTRAINT star_name_length CHECK (LENGTH(name) >= 4 AND LENGTH(name) <= 32),
 	CONSTRAINT star_coordinates UNIQUE (coordinates)
@@ -29,8 +29,8 @@ CREATE TABLE IF NOT EXISTS sales_history (
 	date_sold TIMESTAMP,
 	price_ether NUMERIC(30,18) NOT NULL,
 	star_id BIGINT REFERENCES stars(id) NOT NULL,
-	seller_id INTEGER REFERENCES wallets(id) NOT NULL,
-	buyer_id INTEGER REFERENCES wallets(id) NOT NULL
+	seller_id BIGINT REFERENCES wallets(id) NOT NULL,
+	buyer_id BIGINT REFERENCES wallets(id) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS names_history (
